Create RegressionDataGeneratorBase abstract class
